+++ packages/app/src/api/core-domain/repo/IOwnershipRequestRepo.ts
@@ -1,6 +1,6 @@
 import type { OwnershipRequest } from "@common/core-domain/domain/OwnershipRequest";
 import type { OwnershipRequestStatus } from "@common/core-domain/domain/valueObjects/ownership_request/OwnershipRequestStatus";
-import type { Repo } from "@common/shared-domain";
+import type { BulkRepo } from "@common/shared-domain";
 
 export const OWNERSHIP_REQUEST_SORTABLE_COLS = {
   date: "created_at",
@@ -9,9 +9,8 @@ export const OWNERSHIP_REQUEST_SORTABLE_COLS = {
   déclarant: "email",
 };
 
-export interface IOwnershipRequestRepo extends Repo<OwnershipRequest> {
+export interface IOwnershipRequestRepo extends BulkRepo<OwnershipRequest> {
   countSearch({ siren, status }: { siren?: string; status?: OwnershipRequestStatus }): Promise<number>;
-
   search({
     siren,
     status,
@@ -27,4 +26,6 @@ export interface IOwnershipRequestRepo extends Repo<OwnershipRequest> {
     siren?: string;
     status?: OwnershipRequestStatus;
   }): Promise<OwnershipRequest[]>;
+  updateWithOwnership(item: OwnershipRequest): Promise<void>;
+  updateWithOwnershipBulk(...items: OwnershipRequest[]): Promise<void>;
 }