+++ packages/app/src/common/core-domain/domain/valueObjects/NafCode.ts
@@ -1,11 +1,14 @@
 import { NAF } from "@common/dict";
-import { EnumString } from "@common/shared-domain/domain/valueObjects";
+import { TupleString } from "@common/shared-domain/domain/valueObjects";
 import { Object } from "@common/utils/overload";
+import type { UnknownMapping } from "@common/utils/types";
 
 const NAF_KEYS = Object.keys(NAF);
 
-export class NafCode extends EnumString {
-  constructor(value: string) {
+type Naf = typeof NAF_KEYS[number] | UnknownMapping;
+
+export class NafCode extends TupleString<typeof NAF_KEYS> {
+  constructor(value: Naf) {
     super(value, NAF_KEYS);
   }
 }