+++ packages/app/src/common/core-domain/mappers/ownershipRequestMap.ts
@@ -4,6 +4,7 @@ import { Email, UniqueID } from "@common/shared-domain/domain/valueObjects";
 import type { Objectize } from "@common/utils/types";
 
 import { OwnershipRequest } from "../domain/OwnershipRequest";
+import { ErrorDetail } from "../domain/valueObjects/ownership_request/ErrorDetail";
 import { OwnershipRequestStatus } from "../domain/valueObjects/ownership_request/OwnershipRequestStatus";
 import { Siren } from "../domain/valueObjects/Siren";
 import type { OwnershipRequestDTO } from "../dtos/OwnershipRequestDTO";
@@ -18,7 +19,7 @@ export const ownershipRequestMap: Required<Mapper<OwnershipRequest, OwnershipReq
         askerEmail: new Email(raw.asker_email),
         email: raw.email !== null ? new Email(raw.email) : undefined,
         status: new OwnershipRequestStatus(raw.status),
-        errorDetail: raw.error_detail !== null ? raw.error_detail : undefined,
+        errorDetail: raw.error_detail?.length ? new ErrorDetail(raw.error_detail) : undefined,
       },
       new UniqueID(raw.id),
     );